+++ dallinger/utils.py
@@ -3,41 +3,33 @@ import functools
 import io
 import locale
 import os
+import pkg_resources
 import random
-import redis
+import re
 import shutil
 import string
 import subprocess
 import sys
 import tempfile
 import webbrowser
+
+from faker import Faker
+from hashlib import md5
+from pathlib import Path
 from pkg_resources import get_distribution
+from unicodedata import normalize
 
 try:
     from importlib.metadata import files as files_metadata
 except ImportError:
     from importlib_metadata import files as files_metadata
-from six.moves.urllib.parse import urlparse
 
+from dallinger import db
 from dallinger.config import get_config
 from dallinger.compat import is_command
 
 
-def connect_to_redis(url=None):
-    """Return a connection to Redis.
-
-    If a URL is supplied, it will be used, otherwise an environment variable
-    is checked before falling back to a default.
-
-    Since we are generally running on Heroku, and configuring SSL certificates
-    is challenging, we disable cert requirements on secure connections.
-    """
-    redis_url = url or os.getenv("REDIS_URL", "redis://localhost:6379")
-    connection_args = {"url": redis_url}
-    if urlparse(redis_url).scheme == "rediss":
-        connection_args["ssl_cert_reqs"] = None
-
-    return redis.from_url(**connection_args)
+fake = Faker()
 
 
 def get_base_url():
@@ -364,3 +356,421 @@ def get_editable_dallinger_path():
         if os.path.isfile(egg_link):
             return open(egg_link).read().split()[0]
     return None
+
+
+def setup_experiment(
+    log, debug=True, verbose=False, app=None, exp_config=None, local_checks=True
+):
+    """Checks the experiment's python dependencies, then prepares a temp directory
+    with files merged from the custom experiment and Dallinger.
+
+    The resulting directory includes all the files necessary to deploy to
+    Heroku.
+    """
+
+    # Verify that the Postgres server is running.
+    if local_checks:
+        try:
+            log("Checking your local Postgres database connection...")
+            db.check_connection()
+        except Exception:
+            log("There was a problem connecting!")
+            raise
+
+        # Check that the demo-specific requirements are satisfied.
+        try:
+            with open("requirements.txt", "r") as f:
+                dependencies = [r for r in f.readlines() if r[:3] != "-e "]
+        except (OSError, IOError):
+            dependencies = []
+        pkg_resources.require(dependencies)
+    ensure_constraints_file_presence(os.getcwd())
+    # Generate a unique id for this experiment.
+    from dallinger.experiment import Experiment
+
+    experiment_uid = heroku_app_id = Experiment.make_uuid(app)
+    log("Experiment UID: {}".format(experiment_uid))
+
+    # Load and update the config
+    config = get_config()
+    if not config.ready:
+        config.load()  #
+    if exp_config:
+        config.extend(exp_config)
+
+    # If the user provided an app name, store it. We'll use it as the basis for
+    # the Heroku app ID. We still have a fair amount of ambiguity around what
+    # this value actually represents (it's not used as _only_ the Heroku app ID).
+    if app:
+        heroku_app_id = str(app)
+        log("Using custom Heroku ID root: {}".format(heroku_app_id))
+
+    config.extend(
+        {
+            "id": str(experiment_uid),
+            "heroku_app_id_root": str(heroku_app_id),
+        }
+    )
+
+    if not config.get("dashboard_password", None):
+        config.set("dashboard_password", fake.password(length=20, special_chars=False))
+
+    temp_dir = assemble_experiment_temp_dir(log, config, for_remote=not local_checks)
+    log("Deployment temp directory: {}".format(temp_dir), chevrons=False)
+
+    # Zip up the temporary directory and place it in the cwd.
+    if not debug:
+        log("Freezing the experiment package...")
+        shutil.make_archive(
+            os.path.join(os.getcwd(), "snapshots", heroku_app_id + "-code"),
+            "zip",
+            temp_dir,
+        )
+
+    return (heroku_app_id, temp_dir)
+
+
+def ensure_constraints_file_presence(directory: str):
+    """Looks into the path represented by the string `directory`.
+    Does nothing if a `constraints.txt` file exists there and is
+    newer than a sibling `requirements.txt` file.
+    If it exists but is not up to date a ValueError exception is raised.
+    Otherwise it creates the constraints.txt file based on the
+    contents of the `requirements.txt` file.
+
+    If the `requirements.txt` does not exist one is created with
+    `dallinger` as its only dependency.
+    """
+    constraints_path = Path(directory) / "constraints.txt"
+    requirements_path = Path(directory) / "requirements.txt"
+    if not requirements_path.exists():
+        requirements_path.write_text("dallinger\n")
+    requirements_path_hash = md5(requirements_path.read_bytes()).hexdigest()
+    if constraints_path.exists():
+        if requirements_path_hash in constraints_path.read_text():
+            return
+        else:
+            raise ValueError(
+                "\nChanges detected to requirements.txt: run the command\n    dallinger generate-constraints\nand retry"
+            )
+
+    prev_cwd = os.getcwd()
+    try:
+        os.chdir(directory)
+        compile_info = f"dallinger generate-constraints\n#\n# Compiled from a requirement.txt file with sha: {requirements_path_hash}"
+        check_output(
+            ["pip-compile", "requirements.txt", "-o", "constraints.txt"],
+            env=dict(
+                os.environ,
+                CUSTOM_COMPILE_COMMAND=compile_info,
+            ),
+        )
+    finally:
+        os.chdir(prev_cwd)
+    # Make the path the experiment requirements.txt file relative
+    constraints_contents = constraints_path.read_text()
+    constraints_contents_amended = re.sub(
+        "via -r .*requirements.txt", "via -r requirements.txt", constraints_contents
+    )
+    constraints_path.write_text(constraints_contents_amended)
+
+
+def assemble_experiment_temp_dir(log, config, for_remote=False):
+    """Create a temp directory from which to run an experiment.
+    If for_remote is set to True the preparation includes bundling
+    the local dallinger version if it was installed in editable mode.
+    This is always needed for docker debug and deployment, but not needed for
+    local Heroku debugging.
+
+    The new directory will include:
+    - Copies of custom experiment files which don't match the exclusion policy
+    - Templates and static resources from Dallinger
+    - An export of the loaded configuration
+    - Heroku-specific files (Procile, runtime.txt) from Dallinger
+    - A requirements.txt file with the contents from the constraints.txt file
+      in the experiment (Dallinger should have generated one with pip-compile
+      if needed by the time we reach this code)
+    - A dallinger zip (only if dallinger is installed in editable mode)
+    - A prepare_container.sh script (possibly empty)
+
+    Assumes the experiment root directory is the current working directory.
+
+    Returns the absolute path of the new directory.
+    """
+    exp_id = config.get("id")
+    dst = os.path.join(tempfile.mkdtemp(), exp_id)
+
+    # Copy local experiment files, minus some
+    ExperimentFileSource(os.getcwd()).selective_copy_to(dst)
+
+    # Export the loaded configuration
+    config.write(filter_sensitive=True, directory=dst)
+
+    # Save the experiment id
+    with open(os.path.join(dst, "experiment_id.txt"), "w") as file:
+        file.write(exp_id)
+
+    # Copy Dallinger files
+    dallinger_root = dallinger_package_path()
+    ensure_directory(os.path.join(dst, "static", "scripts"))
+    ensure_directory(os.path.join(dst, "static", "css"))
+    frontend_files = [
+        os.path.join("static", "css", "bootstrap.min.css"),
+        os.path.join("static", "css", "dallinger.css"),
+        os.path.join("static", "css", "dashboard.css"),
+        os.path.join("static", "scripts", "jquery-3.5.1.min.js"),
+        os.path.join("static", "scripts", "popper.min.js"),
+        os.path.join("static", "scripts", "bootstrap.min.js"),
+        os.path.join("static", "scripts", "clipboard.min.js"),
+        os.path.join("static", "scripts", "dallinger2.js"),
+        os.path.join("static", "scripts", "network-monitor.js"),
+        os.path.join("static", "scripts", "reqwest.min.js"),
+        os.path.join("static", "scripts", "require.js"),
+        os.path.join("static", "scripts", "reconnecting-websocket.js"),
+        os.path.join("static", "scripts", "spin.min.js"),
+        os.path.join("static", "scripts", "tracker.js"),
+        os.path.join("static", "scripts", "store+json2.min.js"),
+        os.path.join("templates", "error.html"),
+        os.path.join("templates", "error-complete.html"),
+        os.path.join("templates", "launch.html"),
+        os.path.join("templates", "questionnaire.html"),
+        os.path.join("templates", "exit_recruiter.html"),
+        os.path.join("templates", "exit_recruiter_mturk.html"),
+        os.path.join("templates", "waiting.html"),
+        os.path.join("templates", "login.html"),
+        os.path.join("templates", "dashboard_lifecycle.html"),
+        os.path.join("templates", "dashboard_database.html"),
+        os.path.join("templates", "dashboard_heroku.html"),
+        os.path.join("templates", "dashboard_home.html"),
+        os.path.join("templates", "dashboard_monitor.html"),
+        os.path.join("templates", "dashboard_mturk.html"),
+        os.path.join("static", "robots.txt"),
+    ]
+    frontend_dirs = [os.path.join("templates", "base")]
+    for filename in frontend_files:
+        src = os.path.join(dallinger_root, "frontend", filename)
+        dst_filepath = os.path.join(dst, filename)
+        if not os.path.exists(dst_filepath):
+            shutil.copy(src, dst_filepath)
+    for filename in frontend_dirs:
+        src = os.path.join(dallinger_root, "frontend", filename)
+        dst_filepath = os.path.join(dst, filename)
+        if not os.path.exists(dst_filepath):
+            shutil.copytree(src, dst_filepath)
+
+    # Copy Heroku files
+    heroku_files = ["Procfile"]
+    for filename in heroku_files:
+        src = os.path.join(dallinger_root, "heroku", filename)
+        shutil.copy(src, os.path.join(dst, filename))
+
+    # Write out a runtime.txt file based on configuration
+    pyversion = config.get("heroku_python_version")
+    with open(os.path.join(dst, "runtime.txt"), "w") as file:
+        file.write("python-{}".format(pyversion))
+
+    if not config.get("clock_on"):
+        # If the clock process has been disabled, overwrite the Procfile:
+        src = os.path.join(dallinger_root, "heroku", "Procfile_no_clock")
+        shutil.copy(src, os.path.join(dst, "Procfile"))
+
+    dst_prepare_container = Path(dst) / "prepare_container.sh"
+    if not dst_prepare_container.exists():
+        shutil.copyfile(
+            Path(dallinger_root) / "docker" / "prepare_container.sh",
+            dst_prepare_container,
+        )
+
+    ExplicitFileSource(os.getcwd()).selective_copy_to(dst)
+
+    requirements_path = Path(dst) / "requirements.txt"
+    # Overwrite the requirements.txt file with the contents of the constraints.txt file
+    (Path(dst) / "constraints.txt").replace(requirements_path)
+    if for_remote:
+        dallinger_path = get_editable_dallinger_path()
+        if dallinger_path and not os.environ.get("DALLINGER_NO_EGG_BUILD"):
+            log(
+                "Dallinger is installed as an editable package, "
+                "and so will be copied and deployed in its current state, "
+                "ignoring the dallinger version specified in your experiment's "
+                "requirements.txt file!\n"
+                "If you don't need this you can speed up startup time by setting "
+                "the environment variable DALLINGER_NO_EGG_BUILD:\n"
+                "    export DALLINGER_NO_EGG_BUILD=1\n"
+                "or you can install dallinger without the editable (-e) flag."
+            )
+            egg_name = build_and_place(dallinger_path, dst)
+            # Replace the line about dallinger in requirements.txt so that
+            # it refers to the just generated package
+            constraints_text = requirements_path.read_text()
+            new_constraints_text = re.sub(
+                "dallinger==.*", f"file:{egg_name}", constraints_text
+            )
+            requirements_path.write_text(new_constraints_text)
+    return dst
+
+
+class ExperimentFileSource(object):
+    """Treat an experiment directory as a potential source of files for
+    copying to a temp directory as part of a deployment (debug or otherwise).
+    """
+
+    def __init__(self, root_dir="."):
+        self.root = root_dir
+        self.git = GitClient()
+
+    @property
+    def files(self):
+        """A Set of all files copyable in the source directory, accounting for
+        exclusions.
+        """
+        return set(self._walk())
+
+    @property
+    def size(self):
+        """Combined size of all files, accounting for exclusions."""
+        return sum([os.path.getsize(path) for path in self._walk()])
+
+    def selective_copy_to(self, destination):
+        """Write files from the source directory to another directory, skipping
+        files excluded by the general exclusion_policy, plus any files
+        ignored by git configuration.
+        """
+        for path in self.files:
+            subpath = os.path.relpath(path, start=self.root)
+            target_folder = os.path.join(destination, os.path.dirname(subpath))
+            ensure_directory(target_folder)
+            shutil.copy2(path, target_folder)
+
+    def _walk(self):
+        # The GitClient and os.walk may return different representations of the
+        # same unicode characters, so we use unicodedata.normalize() for
+        # comparisons:
+        # list(name_from_git)
+        # ['å', ' ', 'f', 'i', 'l', 'e', '.', 't', 'x', 't']
+        # list(from_os_walk)
+        # ['a', '̊', ' ', 'f', 'i', 'l', 'e', '.', 't', 'x', 't']
+        exclusions = exclusion_policy()
+        git_files = {
+            os.path.join(self.root, normalize("NFC", f)) for f in self.git.files()
+        }
+        for dirpath, dirnames, filenames in os.walk(self.root, topdown=True):
+            current_exclusions = exclusions(dirpath, os.listdir(dirpath))
+            # Modifying dirnames in-place will prune the subsequent files and
+            # directories visited by os.walk. This is only possible when
+            # topdown = True
+            dirnames[:] = [d for d in dirnames if d not in current_exclusions]
+            legit_files = {
+                os.path.join(dirpath, f)
+                for f in filenames
+                if f not in current_exclusions
+            }
+            if git_files:
+                normalized = {normalize("NFC", str(f)): f for f in legit_files}
+                legit_files = {v for k, v in normalized.items() if k in git_files}
+            for legit in legit_files:
+                yield legit
+
+
+class ExplicitFileSource(object):
+    """Add files that are explicitly requested by the experimenter with a hook function."""
+
+    def __init__(self, root_dir="."):
+        self.root = root_dir
+
+    def _get_mapping(self, dst):
+        from dallinger.config import initialize_experiment_package
+
+        initialize_experiment_package(dst)
+        from dallinger.experiment import load
+
+        exp_class = load()
+        extra_files = getattr(exp_class, "extra_files", None)
+        if extra_files is None:
+            try:
+                from dallinger_experiment.experiment import extra_files
+            except ImportError:
+                try:
+                    from dallinger_experiment.dallinger_experiment import extra_files
+                except ImportError:
+                    pass
+
+        if extra_files is not None:
+            for src, filename in extra_files():
+                filename = filename.lstrip("/")
+                if os.path.isdir(src):
+                    for dirpath, dirnames, filenames in os.walk(src, topdown=True):
+                        for fn in filenames:
+                            dst_fileparts = [dst, filename] + dirnames + [fn]
+                            dst_filepath = os.path.join(*dst_fileparts)
+                            yield (
+                                os.path.join(dirpath, fn),
+                                dst_filepath,
+                            )
+                else:
+                    dst_filepath = os.path.join(dst, filename)
+                    yield (src, dst_filepath)
+
+    @property
+    def files(self):
+        """A Set of all files copyable in the source directory, accounting for
+        exclusions.
+        """
+        return {src for (src, dst) in self._get_mapping("")}
+
+    @property
+    def size(self):
+        """Combined size of all files, accounting for exclusions."""
+        return sum([os.path.getsize(path) for path in self.files])
+
+    def selective_copy_to(self, destination):
+        """Write files declared in extra_files from the source directory
+        to another directory.
+        """
+        for from_path, to_path in self._get_mapping(destination):
+            target_folder = os.path.dirname(to_path)
+            ensure_directory(target_folder)
+            shutil.copyfile(from_path, to_path)
+
+
+def exclusion_policy():
+    """Returns a callable which, when passed a directory path and a list
+    of files in that directory, will return a subset of the files which should
+    be excluded from a copy or some other action.
+
+    See https://docs.python.org/3/library/shutil.html#shutil.ignore_patterns
+    """
+    patterns = set(
+        [
+            ".git",
+            "config.txt",
+            "*.db",
+            "*.dmg",
+            "node_modules",
+            "snapshots",
+            "data",
+            "server.log",
+            "__pycache__",
+        ]
+    )
+
+    return shutil.ignore_patterns(*patterns)
+
+
+def build_and_place(source: str, destination: str) -> str:
+    """Builds a python egg with the source found at `source` and places it in
+    `destination`.
+    Only works if dallinger is currently installed in editable mode.
+
+    Returns the full path of the newly created distribution file.
+    """
+    old_dir = os.getcwd()
+    try:
+        os.chdir(source)
+        check_output(["python", "-m", "build"])
+        # The built package is the last addition to the `dist` directory
+        package_path = max(Path(source).glob("dist/*"), key=os.path.getctime)
+        shutil.copy(package_path, destination)
+    finally:
+        os.chdir(old_dir)
+    return package_path.name
